Extract annotation fixtures into a separate module
Also make all of the fixtures pure functions and move
the logic for setting up the drafts fake into the
createDirective() function in the annotation tests.
